projects/telemetry_cli/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "telemetry-cli"
version = "0.1.0"
description = "Universal telemetry agent (real sources + synthetic signals) -> Experiment Service telemetry ingest."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "product_dev_course" }]
dependencies = [
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "aiohttp>=3.10.0",
]

[project.scripts]
telemetry-cli = "telemetry_cli.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/telemetry_cli"]


